+++ apps/docs/content/3-components/2-library/text-input-password.mdx
@@ -1,0 +1,97 @@
+---
+title: Text Input Password
+description: A password input field with visibility toggle functionality.
+status: in-review
+libraries:
+  - platform: 'figma'
+    link: 'https://www.figma.com/design/ZEqfpXhocdZs1K5YfsE5ws/Design-system?node-id=4008-3093&t=x9tccZ94Zz8w2Tvu-4'
+    status: 'alpha'
+  - platform: 'global'
+    link: ''
+    status: 'alpha'
+  - platform: react
+    link: ''
+    status: 'alpha'
+---
+
+# Text Input Password
+
+The `TextInputPassword` component provides a secure password input field with an integrated visibility toggle button. This allows users to switch between masked (`password`) and unmasked (`text`) input modes.
+
+## Variants
+
+### Default
+
+<ComponentContainer>
+  <FormField
+    hint={{
+      text: 'Hint: Enter your password securely.',
+    }}
+    label={{
+      htmlFor: 'password-input-id',
+      text: 'Password',
+    }}
+  >
+    <TextInputPassword data-testid="password-input-id" id="password-input-id" />
+  </FormField>
+</ComponentContainer>
+
+<Tabs id="tab3">
+  <TabList>
+    <TabItem value="tab31" checked>HTML</TabItem>
+    <TabItem value="tab33">React</TabItem>
+  </TabList>
+  <TabPanel value="tab31">        
+    ```html
+    <fieldset class="">
+    <label class="gi-text-md gi-label gi-mb-1" for="password-input-id">Password</label>
+        <div class="gi-hint-text-md gi-hint-text gi-mb-1">Hint: Enter your password securely.</div>
+        <div class="gi-text-input-container">
+            <div class="gi-text-input-inner">
+                <input data-icon-start="false" data-icon-end="false" data-end-element="true" data-prefix="false" data-suffix="false" class="gi-text-input" data-testid="password-input-id" id="password-input-id" type="password">
+                <div class="gi-text-input-end-element" data-suffix="false"><button role="button" data-testid="govieIconButton-dark-flat-small-undefined" class="gi-btn gi-btn-flat-dark gi-icon-btn-small undefined "><span data-testid="govie-icon" role="presentation" class="gi-block material-symbols-outlined" style="font-size: 16px;">visibility</span></button></div>
+            </div>
+        </div>
+    </fieldset>
+    ```
+  </TabPanel>
+  <TabPanel value="tab33">
+    ```jsx
+    import { FormField, TextInputPassword } from "@govie-ds/react";
+    
+    <FormField
+      hint={{
+        text: 'Hint: Enter your password securely.'
+      }}
+      label={{
+        htmlFor: 'password-input-id',
+        text: 'Password'
+      }}
+    >
+      <TextInputPassword
+        data-testid="password-input-id"
+        id="password-input-id"
+      />
+    </FormField>
+    ```
+  </TabPanel>
+</Tabs>
+
+## Usage
+
+### Best Practices
+
+### When to Use This Component
+
+- **Secure Input Fields**: Use this component for password fields in login, signup, and authentication forms.
+- **Toggleable Password Visibility**: Ideal for user convenience, allowing users to verify their password input.
+- **Error Handling**: Works well with validation messages for incorrect passwords.
+- **Accessibility Support**: Includes `aria-label` for visibility toggling to enhance screen reader usability.
+
+### When Not to Use This Component
+
+- **Non-Sensitive Inputs**: If the field does not contain sensitive information, use a standard `TextInput` instead.
+
+## Status
+
+<ComponentStatusBlock componentId="text-input-password" />
