+++ packages/react/ds/src/index.ts
@@ -116,12 +116,11 @@ export { TabList } from './tabs/tab-list.js';
 export { TabPanel } from './tabs/tab-panel.js';
 export { Tabs, type TabsProps } from './tabs/tabs.js';
 export { Tag, type TagProps, TagTypeEnum, type TagType } from './tag/tag.js';
+export { TextInput, type TextInputProps } from './text-input/text-input.js';
 export {
-  TextInput,
   TextInputPassword,
   type TextInputPasswordProps,
-  type TextInputProps,
-} from './text-input/text-input.js';
+} from './text-input-password/text-input-password.js';
 export { TextArea, type TextAreaProps } from './textarea/textarea.js';
 export { toaster, ToastProvider } from './toast/toast.js';
 export type {
